refactor: split up FoundationError into method-specific errors
