+++ src/dfx-core/src/error/get_current_exe.rs
@@ -1,0 +1,7 @@
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum GetCurrentExeError {
+    #[error("Failed to identify currently running executable: {0}")]
+    NoCurrentExe(std::io::Error),
+}
