+++ src/dfx-core/src/error/get_user_home.rs
@@ -1,10 +1,7 @@
 use thiserror::Error;
 
 #[derive(Error, Debug)]
-pub enum FoundationError {
+pub enum GetUserHomeError {
     #[error("Cannot find home directory (no HOME environment variable).")]
     NoHomeInEnvironment(),
-
-    #[error("Failed to identify currently running executable: {0}")]
-    NoCurrentExe(std::io::Error),
 }
