+++ src/dfx-core/src/error/identity/get_legacy_credentials_pem_path.rs
@@ -1,8 +1,8 @@
-use crate::error::foundation::FoundationError;
+use crate::error::get_user_home::GetUserHomeError;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
 pub enum GetLegacyCredentialsPemPathError {
     #[error("Failed to get legacy pem path: {0}")]
-    GetLegacyPemPathFailed(FoundationError),
+    GetLegacyPemPathFailed(GetUserHomeError),
 }
