+++ src/dfx-core/src/error/mod.rs
@@ -7,8 +7,9 @@ pub mod config;
 pub mod dfx_config;
 pub mod encryption;
 pub mod extension;
-pub mod foundation;
 pub mod fs;
+pub mod get_current_exe;
+pub mod get_user_home;
 pub mod identity;
 pub mod keyring;
 pub mod load_dfx_config;
